move plugin compatibility code to separate files, add support for The Events Calendar to automatically show description on List/Month/Day/etc. pages (Closes #17), add extra check to prevent errors on The Events Calendar Event Category pages (fixes #18)
